Implement if-modified-since correctly for ESI
We don't actually know the last-modified value of an ESI 
document until we have assembled it. At this point it is
too late to give a correct value.

This patch starts off by giving the first requestor
of an object the last-modified of the container. 
This is safe because the worse that can happen is that
a fragment has a higher last-modified and we waste
some bandwidth. (Boy, trailers would be nice)

As we assemble the fragments, we keep track of the highest
last-modified value and store that in the object. 
Next time the object is requested, we return that 
last-modified.

If we get an IMS and we have never assembled the object
we assemble the object and then return it. 
(A future optimisation is to assemble it and then 
send not-modified if it isn't, purely to save bandwidth)

If we have a cached value then we base the IMS of it.

There is a race here that if a fragment is updated
and noone has requested the entire object the cached
value is stale and we will return not modified.

I think this is fine for now since otherwise we have to
a) keep reverse dependencies or
b) pull all fragments together if it isn't needed

A user can always get the latest version by doing
a force reload, since the browser doesn't send 
if-modified-since on those.

An improvement would be to keep track of the time
any fragment expires an only cache the last-modified
that long. Which would solve the problem for anyone
who doesn't purge fragments to get them to update.

This should probably have a parameter



git-svn-id: http://www.varnish-cache.org/svn/branches/sky@4119 d4fa192b-c00b-0410-8231-f00ffab90ce4