Added -b (buffer size) parameter to varnishtest
When implementing in VCL complex logic (not necessarely related with
caching contents -e.g. access control, rate limiting, etc.-) varnishtest
is a powerfull tool to automate testing. These type of tests sometimes
are complex and they are not easy to split is smaller pieces.

In some extreme cases, the varnishlog buffer space allocated by
varnishtest is not enough when testing complex VCL logic including a
log of logging. The size of that buffer is hard coded and it has been
increased  in the past [1]. This simple commit allows selecting the size
of that buffer in the command line.

[1] https://www.varnish-cache.org/trac/changeset/9511eefcdafea9a6c87f83a46d5e2b1d41c1c2c9