build: support for asan/tsan/msan/ubsan and gcov
Modern versions of gcc and clang support various runtime sanitizers that
can be used to find nasty bugs at runtime. Including support for these
in the build is useful, but also it is useful to know how much code
coverage you're getting with tests when using these sanitizers.

This patch adds build-time support for the sanitizers as well as gcov
support to get code coverage metrics on varnish tests. To get HTML
output of this information, "lcov" is a useful tool. The general
workflow is:

	make
	lcov -o varnish_initial.info -i -c -d `pwd`
	make check
	lcov -o varnish_test.info -c -d `pwd`
	lcov -a varnish_initial.info -a varnish_test.info -o varnish_total.info

HTML output can then be generated using lcov's genhtml tool. This
information can then be evaluated to determine whether more tests are
needed to provide more test-time safety guarantees based on an initial
lack of warnings or errors from sanitizer tools.