docs: Add FP-math weirdness caveats on 32bit intel
The overhead required to debug this and then to find a proper solution for
broken compilers warrants a documentation change. It may be possible to "be
smart" in configure, but I am also wary of changing FP compiles for the
numerous versions of broken compilers that currently don't produce broken
code (as far as we know).

Thanks to Ingvar Hagelund for finding this issue and being patient and
helpful in debugging it.